:recycle: relocate authentication strategies
